Determine how to define schema.condelists
This is to be a set. This data structure manages uniqueness
itself, so there is no need to do this manually. To make
uniqueness be checked properly, a custom hash function for
Codelists will be required.

It is not necessary abstract the underlying data structure.

Any side-effects can be managed in the future using either
a subclass of set, improved hash functions, or a combination
of both of these things.