Flatten structure of the codelists folder
The version-specific codelist folders have been flattened, and now contain all codelists
that can be used by that version of the Standard: embedded and non-embedded.  Non-embedded
codelists are included through the use of symlinks, which create a link to the
2xx/codelists_non_embedded folder.

Following this, approproate changes have been made to code and tests.