[#24] Add logic to only convert strings to bytes in python3
Adds additional logic statement to check python version. This is
necessary as I couldn't find a way to make the underlaying
functionality compatible with python2 and 3 due to the differences
in the way both versions handle strings - i.e. strings in python2
are treated as bytes objects by default, so converting to bytes is
not necessary in the first place.