Add pip requirement - 'future'
This is required to provide the same functionality of 'math.ceil()' on 2.7 as 3+
At Python 2, this function returns a float, while at v3 it returns an int.

The range() function requires an integer to be passed. As such, the options are
either this or casting an an int(). This method makes it clearer what will need
changing at such as point as 2.7 compatibility is dropped.