Categorise XMLSyntaxErrors
This converts the lxml errors into something IATI-specific.

Should there not be a specific case for a particular error, a
generic error will be provided. In an actual implementation, any
errors that do not have a specific case should be logged to help
identify errors that occur frequently but have not been categorised.