Prevent error buildup
It appears that lxml contains an internal error log that builds up over time.
When a test function is parameterised by pytest, all calls to the function
add to this singular error log. As such, later calls to the function will be
provided with the errors from earlier calls. This is bad.

By providing a custom XMLParser every time, the error log only refers to a single
call to a test function. As such, buildup of past errors does not occur and the
tests are checking the correct information. This is good.