Reduce the range of tested encoding mismatches
When there is an encoding mismatch it would be expected that the
ERR_INVALID_ENCODING error is raised. This is not always the case.

Other errors including ERR_INVALID_CHAR, ERR_GT_REQUIRED and
ERR_DOCUMENT_EMPTY may also occur depending on the document being
processed. Something like a newline can completely change the
errors that are output.

The XML spec permits a full range of character encodings to be
used. libxml2, however, only supports UTF-8, UTF-16, ASCII and
ISO-Latin-1 by default. Even within this set of supported
encodings, the expected errors are not returned consistently
(if at all).

With this in mind, a limited set of encodings will be tested.
Other encodings will lead to somewhat unhelpful errors being
returned. Since the probability of these situations occurring
is fairly low, this shouldn't be a major problem.