Raise validation errors when given an empty string
An empty string is not parsable XML, so should be treated as invalid.
Upon creation of a Dataset, an error may be raised when creating an
etree from an empty string despite the validator not having any
problem with an exuivalent situation. As such, the empty string is
treated as a special case.