Move validation error codes into a function
With loading the codes from a file, there ends up being a
reference to iati.core at the module-level. Due to the
shorthand-imports, iati.validator is called via iati.core.data
in the iati.core __init__.py file.

At this point, iati.core is not defined and so the reference
failed. Putting it in a function makes this work.