Make date_order rule work in 2.7
Python2.7 does not have a fullmatch function for regexes.
This is resolved by setting start and end anchors in the
regex of interest.

Having fixed the regex problem, it becomes apparent that
there is another problem in that None is orderable at v2
but not v3. This is also fixed.