Remove TODOs section from README
Leaving this here adds additional maintenance burden, and
this is not undertaken.

The Issues are used for this purpose, and are able to adapt
over time, with discussions and additional information being
added.

The current situation of using Issues isn't ideal, though
it's not immediately obvious what a correct solution is.