Update the changelog for #45
This is most of the information, though there may be a few
small things missing.

The issue references are also not directly obvious in all
situations - there is basic information, though the references
are not gospel for all the changes in a point, and may also
cover a range of other changes.

This is due to the questionable branching (sorry...) while
implementing this functionality. Learning has been done and
things should be better in the future.