Test loading of strings with various encodings
This tests all of the encodings present in published IATI data, plus UTF-32.

Testing of UTF encodings is done by checking for the presence of characters
that are different numbers of bytes in each of the encodings.

windows-1252 is checked by looking for a character that has is different to
ASCII - many of its characters are the same.

Files were generated using 'iconv'.
The BOMs were added to the UTF-16[L|B]E files afterwards so that they are
correctly detected by chardet. This was done by following the instructions
at https://superuser.com/a/233113