Rename version-specific resource folders
Fixes #185

Up to this point  folders for specific versions are named, eg. ,  and .

By doing this, it is ambiguous where in the folder name the integer part ends and the decimal part begins. This will become even more of a problem should SemVer be used.As such, there should be a separator between the components of the Version Number.

A hyphen is chosen since it is a clear separator and has no special meaning in either regular expressions (outside of [character classes](https://stackoverflow.com/questions/9589074/regex-should-hyphens-be-escaped), which is a non-issue) or in filepaths on any sane filesystem.