Handle element text() needing to be valid against a Codelist
This is a fix for #224

The Codelist README states the following in relation to mapping files:

  It's structured as a list of mapping elements, which each have a path element that describes the relevant attribute

This indicates that it's incorrect that element text should match a Codelist. IATI, however, is special. This means that at the moment this is deemed something the Standard does.

Until the restriction is removed, this fix is required.

Yay, IATIland...