Make code extraction DRYer
As suggested in:
  https://github.com/IATI/pyIATI/pull/229#pullrequestreview-75269366

Passing strings (rather than lists of strings) around is clearer. It also means that a couple of variables only need defining once.

It is fine that the attrib and text extraction functions take different numbers of parameters since the former clearly needs more information to undertake its task than the latter.