Move version functionality into version module
Previously it was necessary to have a couple of unpleasant import guards when dealing with the constants module.
Functionality relating to versions that was previously in the constants and utility modules is being brought into the version modules.

Completing this fully means that common functionality is now together in a central location.

There will be some further work required to ensure that this migration has been completed in a full and proper manner.