Change constant meaning 'version independent' to not None
This allows '==' to be used to check whether a value is this or not, rather than requiring use of 'is'.
By doing this, it gives more scope to change the value in the future.

A constant has been added to contain this value, and should be used if this meaning is required.

For original explanation for change, see:
    https://github.com/IATI/pyIATI/issues/218#issuecomment-364086162