standardise -> normalise
RE: https://github.com/IATI/pyIATI/pull/280#discussion_r171539607

It is noted that the word 'standardise' may imply that something is being converted to a format that meets the IATI Standard.
This is not always the case.

In the modified cases, 'normalise' has a similar meaning, while not having this potential unclear meaning.
It should be clear from context that 'normalisation' in this context is not about databases, and instead about something else.