Mark functions to update when adding a new version
When adding a new known version of the IATI Standard, a number of tests need changing.
This ensures that errors are raised if a test is said to have been updated for the latest version of the Standard, but notes that the latest version is older than it actually is.

A slightly clunky method to do this is used because fixtures cannot be marked in the same manner as test functions.