Change to use a more appropriate method of marking latest versions
A convoluted method of marking was used in a64197378f2ab990b1399d7531df0ac09d80e7ec - this was to provide a way to mark fixtures.

A more thorough examination of the docs located https://docs.pytest.org/en/latest/builtin.html#_pytest.fixtures.FixtureRequest
This provides a manner in which markers should be added to fixtures.
This commit switches the marking method to use this improved method of marking.