Start full redesign of the resource module tests
This is a process starting with the lowest level components, and building up to the larger components.

Since the resources module was originally created, it has been redesigned multiple times. Across this period, the tests have not been rearchitectured.
This process will turn the resource tests into a state that works for the current module state, rather than that from a year ago.