Test creation of paths to Ruleset resources
This tests valid and invalid paths and versions.
It tests major, minor and version-independent versions.
It tests TypeErrors and ValueErrors

There is access to a private function within 'create_ruleset_path'. This is due to decorators expecting the version argument to be first, and the difficulty in changing this in pre-3.5 versions of Python.
See #294 for more info.