More thoroughly test iati.resources.get_ruleset_paths()
This adds tests to check the various outputs that may exist.
The current implementation is a bit clunky and could be simplified if there were is_major() and is_minor() functions in the version module.
The returned paths are guaranteed to lead to a file at the point at which they are returned.