Remove default values for resource version args
A number of functions in the resources module still had default values for the version argument.
Explicit is better than implicit. As such, the default values have been removed and tests added to ensure TypeErrors are raised as appropriate.

It was documented in the changelog that this was already the case, though the removal of defaults had only been undertaken for some functions.